+++ src/game/logic/pausemenumodel/step.cpp
@@ -1,6 +1,6 @@
 #include "pausemenumodel.ih"
 
-void PauseMenuModel::step(float deltaTime)
+void PauseMenuModel::step(double deltaTime)
 {
   // No need to do anything for the pauser.
 }
